refactor: declare lazy screens at module scope behind one `Suspense` per route switch
