--- apps/desktop/src/mobile/screens/graphs-dynamic.tsx
@@ -1,14 +1,0 @@
-import { lazy, Suspense, type ReactElement } from 'react'
-import { LoadingScreen } from '@/components/loading-screen'
-
-const MobileGraphs = lazy(() =>
-  import('@/mobile/screens/graphs').then((module) => ({ default: module.MobileGraphs })),
-)
-
-export function MobileGraphsDynamic(): ReactElement {
-  return (
-    <Suspense fallback={<LoadingScreen />}>
-      <MobileGraphs />
-    </Suspense>
-  )
-}
