--- apps/desktop/src/mobile/screens/tasks-dynamic.tsx
@@ -1,14 +1,0 @@
-import { lazy, Suspense, type ReactElement } from 'react'
-import { LoadingScreen } from '@/components/loading-screen'
-
-const MobileTasks = lazy(() =>
-  import('@/mobile/screens/tasks').then((module) => ({ default: module.MobileTasks })),
-)
-
-export function MobileTasksDynamic(): ReactElement {
-  return (
-    <Suspense fallback={<LoadingScreen />}>
-      <MobileTasks />
-    </Suspense>
-  )
-}
